feat: add auth for some commentary endpoints

+++ api_v1/commentaries/views.py
@@ -1,6 +1,6 @@
 from typing import Annotated
 
-from fastapi import APIRouter, Depends, status, Path
+from fastapi import APIRouter, Depends, status, Path, HTTPException
 from sqlalchemy.ext.asyncio import AsyncSession
 
 from core.models import db_helper
@@ -10,6 +10,7 @@ from api_v1.commentaries.schemas import (
     CommentaryCreate,
 )
 from api_v1.commentaries import dependencies
+from api_v1.auth.dependencies import get_current_user
 
 router = APIRouter(prefix="/commentaries", tags=["commentaries"])
 
@@ -20,10 +21,14 @@ router = APIRouter(prefix="/commentaries", tags=["commentaries"])
 async def create_commentary(
     comment: CommentaryCreate,
     post_id: Annotated[int, Path],
+    user_id: int = Depends(get_current_user),
     session: AsyncSession = Depends(db_helper.session_dependency),
 ):
     return await crud.create_commentary(
-        session=session, comment_to_create=comment, post_id=post_id
+        session=session,
+        comment_to_create=comment,
+        post_id=post_id,
+        user_id=user_id,
     )
 
 
@@ -37,9 +42,14 @@ async def get_commentary(
 @router.delete("/{comment_id}", status_code=status.HTTP_204_NO_CONTENT)
 async def delete_commentary(
     commentary: Commentary = Depends(dependencies.get_comment_by_id),
+    user_id: int = Depends(get_current_user),
     session: AsyncSession = Depends(db_helper.session_dependency),
 ):
-    return await crud.delete_commentary(
-        session=session,
-        commentary_db=commentary,
-    )
+    if commentary.user_id == user_id:
+
+        return await crud.delete_commentary(
+            session=session,
+            commentary_db=commentary,
+        )
+
+    raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="It is not your commentary")
